+++ src/main/java/tn/esprit/ecocycletech/Entity/AppareilsManagement/Facture.java
@@ -18,7 +18,7 @@ import java.util.Date;
 public class Facture implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long idFacture;
+    private int idFacture;
     @Temporal(TemporalType.DATE)
     private Date date;
     @OneToOne
